Reduce work being done in the executor.
Since I can't/won't be unit testing that, I want it to do as little as
possible. It's a controllable front for the Kubernetes API actions, and
that's it. Determination of service status and such should go somewhere
that I can test it, even if that means dealing with more Kubernetes API
structs in my tests.