Fix redeploymeng issues for alias-linked services.
The service-name metadata – which refers to the name of the Panamax
service - was not being stored with the correct name for aliased links.
Panamax services become ReplicationControllers, which may spawn 1 or
more K8s services (one for each link, plus one for each alias). The
"service-name" metadata is used to link the service to a
ReplicationController to clean things up when a Panamax service is
deleted. The alias services were not being cleaned up, and so
redeployment was failing trying to re-create an existing service.